+++ src/test/java/study/Formula.java
@@ -1,13 +1,13 @@
 package study;
 
 public class Formula {
-    private String formula;
+    private String[] formula;
 
     public Formula(String formula) {
-        this.formula = formula;
+        this.formula = formula.split(" ");
     }
 
-    String[] splitFormula() {
-        return formula.split(" ");
+    public String[] getFormula() {
+        return formula;
     }
 }
