+++ src/test/java/study/Main.java
@@ -8,8 +8,7 @@ public class Main {
         System.out.println("문자열 계산기");
         System.out.println("공백으로 구분하여 원하시는 숫자와 연산자를 입력해주세요.");
         String value = scanner.nextLine();//사용자로부터 공백을 기준으로 문자열을 구분하여 입력 받음
-        String[] formula = new Formula(value).splitFormula();
         StringCalculator stringCalculator = new StringCalculator();
-        System.out.println(stringCalculator.calculateFormula(formula));
+        System.out.println(stringCalculator.calculateFormula(new Formula(value).getFormula()));
     }
 }
